create payment and update calculate revenueByYear

+++ src/main/java/com/example/vinhomeproject/dto/ContractDTO_3.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 public class ContractDTO_3 {
     private LocalDate createDateContract;
     private LocalDate expireDateContract;
-    private int durationMonth;
+    private long durationMonth;
     private double totalPrice;
     private double pricePerMonth;
     private String areaName;
